fix compiler error on mac osx %llu format

+++ sharedlib.c
@@ -195,8 +195,8 @@ static size_t process_nums(const char *ops) {
 
     if (val <= 0) {
 
-        fprintf(stderr, "%s: Invalid specified shared memory size : ringbuffer size <= %lli \n"
-                "Usage: %s [-m] <ringbuffer elements>\n", programmName, SIZE_MAX /sizeof(int), programmName);
+        fprintf(stderr, "%s: Invalid specified shared memory size : ringbuffer size <= %zu \n"
+                "Usage: %s [-m] <ringbuffer elements>\n", programmName,(size_t) (SIZE_MAX /sizeof(int)), programmName);
 
         exit(EXIT_FAILURE);
     } else if (*end != '\0') {
